+++ lib/flutter_password_input.dart
@@ -1,2 +1,4 @@
+export 'package:just_tooltip/just_tooltip.dart' show TooltipAnimation;
+
 export 'src/password_text_field.dart';
 export 'src/password_text_field_theme.dart';